requirements.txt
asgiref==3.3.1
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
coverage==6.4.1
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.6
django-cors-headers==3.11.0
django-crispy-forms==1.11.1
django-filter==2.4.0
djangorestframework==3.13.0
docker==4.3.1
dockerpty==0.4.1
docopt==0.6.2
flake8==4.0.1
gitdb==4.0.7
GitPython==3.1.7
google-auth==1.31.0
httpie==2.6.0
idna==2.10
ipdb==0.13.4
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
kubernetes==11.0.0
mccabe==0.6.1
oauthlib==3.1.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.11
prompt-toolkit==3.0.16
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycountry==22.3.5
pyflakes==2.4.0
Pygments==2.8.0
pykwalify==1.7.0
PySocks==1.7.1
python-box==5.1.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
semver==2.13.0
six==1.16.0
smmap==4.0.0
soupsieve==2.2.1
spython==0.0.79
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
websocket-client==1.1.0
yarg==0.1.9
